extract command logic into utility
